+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Added new resource SPWebApplicationExtension
 * Updated documentation in regards to guidance on installing binaries from
   network locations instead of locally
 * New resources: SPFarmPropertyBag